pyproject.toml
[tool.poetry]
name = "faq2html"
version = "0.2.0"
description = "Convert a structured docx file with FAQs to html."
authors = ["Abhishek Basu <abhishek.sfy@gmail.com>"]
readme = "README.md"
license = "MIT"

[tool.poetry.dependencies]
python = "^3.7"
dominate = "^2.7.0"
click = "^8.1.3"
python-docx = "^0.8.11"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
